+++ packages/caip/src/adapters/coingecko/utils.ts
@@ -35,16 +35,6 @@ export type CoingeckoCoin = {
 
 type AssetMap = Record<ChainId, Record<AssetId, string>>
 
-export const writeFiles = async (data: AssetMap) => {
-  await Promise.all(
-    Object.entries(data).map(async ([chainId, assets]) => {
-      const path = `./src/adapters/coingecko/generated/${chainId}/adapter.json`.replace(':', '_')
-      await fs.promises.writeFile(path, JSON.stringify(assets))
-    })
-  )
-  console.info('Generated CoinGecko AssetId adapter data.')
-}
-
 export const fetchData = async (URL: string) => (await axios.get<CoingeckoCoin[]>(URL)).data
 
 export const parseData = (coins: CoingeckoCoin[]): AssetMap => {
@@ -95,4 +85,14 @@ export const parseData = (coins: CoingeckoCoin[]): AssetMap => {
     [cosmosChainId]: cosmosAssetMap,
     [osmosisChainId]: osmosisAssetMap
   }
+}
+
+export const writeFiles = async (data: AssetMap) => {
+  await Promise.all(
+    Object.entries(data).map(async ([chainId, assets]) => {
+      const path = `./src/adapters/coingecko/generated/${chainId}/adapter.json`.replace(':', '_')
+      await fs.promises.writeFile(path, JSON.stringify(assets))
+    })
+  )
+  console.info('Generated CoinGecko AssetId adapter data.')
 }